Delivery API : - BusinessSummary API updated

+++ src/main/java/com/ahold/technl/sandbox/dto/BusinessSummary.java
@@ -1,4 +1,4 @@
 package com.ahold.technl.sandbox.dto;
 
-public record BusinessSummary(int deliveries,int averageMinutesBetweenDeliveryStart) {
+public record BusinessSummary(int deliveries,long averageMinutesBetweenDeliveryStart) {
 }
